+++ App/Indoor-React-Native/app.json
@@ -4,12 +4,12 @@
     "slug": "Indoor-React-Native",
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "./assets/icon.png",
-    "userInterfaceStyle": "light",
+    "icon": "./assets/leaf.png",
+    "userInterfaceStyle": "dark",
     "splash": {
-      "image": "./assets/splash.png",
+      "image": "./assets/leaf.png",
       "resizeMode": "contain",
-      "backgroundColor": "#ffffff"
+      "backgroundColor": "#000000"
     },
     "assetBundlePatterns": [
       "**/*"
@@ -18,8 +18,11 @@
       "supportsTablet": true
     },
     "android": {
+      "splash": {
+        "resizeMode": "native"
+      },
       "adaptiveIcon": {
-        "foregroundImage": "./assets/adaptive-icon.png",
+        "foregroundImage": "./assets/leaf.png",
         "backgroundColor": "#ffffff"
       }
     },
