+++ App/Indoor-React-Native/src/styles/pagesStyles.jsx
@@ -2,13 +2,13 @@ import { StyleSheet } from "react-native";
 
 export const routesStyles = StyleSheet.create({
   container: {
-    backgroundColor: '#25DFC9',
+    backgroundColor: '#090909',
     flex: 1,
     justifyContent: "center",
     alignItems: "center"
   },
   text: {
-    color: '#000',
+    color: '#fff',
     fontSize: 50
   }
 });
